aws/lambda AppConfig support (#15)
* Appconfig first pass complete

But networkign is the issue - it tries to hit the ipv4 address. To fix this, we need either a nat gateway (solves it generically) or a VPC endpoint per AWS service we're trying to use. The VPC endpoint is cheaper, so i'm planning to go with that for now.

* Turned out the problem was ipv6 disabled on the lambda!

Still found a better pattern for transmitting subnet info though.